requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mlxtend
python-dotenv
SQLAlchemy
psycopg2-binary
networkx
